refactored ProcessFileSystemAttributes

+++ bundles/sirix-core/src/main/java/org/sirix/service/xml/serialize/AbstractSerializer.java
@@ -128,7 +128,7 @@ public abstract class AbstractSerializer implements Callable<Void> {
 
         final int nrOfRevisions = mRevisions.length;
         final int length = (nrOfRevisions == 1 && mRevisions[0] < 0)
-                ? (int) mResMgr.getMostRecentRevisionNumber()
+                ? mResMgr.getMostRecentRevisionNumber()
                 : nrOfRevisions;
 
         for (int i = 1; i <= length; i++) {
